requirements.txt
flask
gunicorn
supabase-py
pycryptodome
